Add an Additional Step on_failure
We want on_failure to both send an email notification and upload the test report file.

Adds another S3 resource called test-report pointing to the test-reports bucket, and a regexp of junit-(.*).xml

Adds an output to the test task with a name of test-report

Updates pipeline configuration to both send the email notification and upload files matching the pattern test-report/junit-*.xml to the test-report resource

We could either use `do` or in `in_parallel`.