Add a Job with an inline Task
In order for the code to be available to our task, we’ll need to add a `get` of the resource to the job `plan`. A task can take inputs from any resource that was referenced by a `get` or a `put` earlier on in the plan.

Inline task = task defined in the pipeline